+++ cmd/tx_send.go
@@ -101,7 +101,7 @@ func BuildCmdTxSend(w io.Writer, handler SendTxHandler, rf *RootFlags) *cobra.Co
 	)
 
 	autoCompleteNetwork(cmd, rf.Home)
-
+	autoCompleteLogLevel(cmd)
 	return cmd
 }