+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### 🛠  Improvements
 - [508](https://github.com/vegaprotocol/vegawallet/issues/508) - Introduce a command to list the service endpoints
 - [507](https://github.com/vegaprotocol/vegawallet/issues/507) - Redirect logs from command service run to a file
+- [516](https://github.com/vegaprotocol/vegawallet/issues/516) - Update protos for market update proposal
 
 ### 🔥 Removal
 - [](https://github.com/vegaprotocol/vegawallet/pull/) -